% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_2.R
\name{hello_2}
\alias{hello_2}
\title{Print a Greeting}
\usage{
hello_2(name)
}
\arguments{
\item{name}{character, name of person to whom greeting should be directed}
}
\value{
nothing useful is returned.
}
\description{
Print a greeting for a custom name
}
\details{
This function make a plot with a greeting to the name passed as an argument to the function
}
\examples{
hello("Chris")

}
